As group - Iteration 1, without exceptions and Junit tests
