--- Code/Part_1/orderlist/src/orderList/menuItems
@@ -1,3 +1,0 @@
-O001 : C001 :COF101,TEA201
-O002 : C002 :COF101,TEA201
-O003 : C003 :COF101,TEA201
